Store list of vertices on stack or queue
Previously, each individual vertex was added to stack or queue. Now the vertex's list of adjacent vertices are added instead.
This allows the program to be run without increasing heap size.
Also refactored DFS and BFS into a single method - will decide later if this is a good idea or not.