Change how vertices store adjacent vertices
It used to be that each vertex would store all of its edges, and the adjacent vertices were retrieved from getting the other vertex from the edges. Instead of saving the edges into a Linked List, we are now saving the adjacent vertices themselves into a Linked List. This saves space because each edge has 2 vertex objects.